# swagger/centralia.yml
openapi: 3.0.3
info:
    description: Service definition for Centralia web services
    version: 1.0.0
    title: Swagger Centralia
    license:
        name: Apache 2.0
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
    - url: "http://localhost:8000"
...
paths:
...
    /usuarios:
        get:
            tags:
                - Usuarios
            summary: Obtener todos los usuarios
            description: Obtiene la lista de usuarios de la plataforma
            operationId: getUser
            responses:
                "200":
                    description: Operacion exitosa
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/RespuestaComun"
                                    - properties:
                                          data:
                                              $ref: "#/components/schemas/UsuariosRespuesta"
                "404":
                    description: Operacion no encontrada
            security:
                - bearerAuth: []
        post:
            tags:
                - Usuarios
            summary: Crear un usuario
            description: Se crea un usuario nuevo
            operationId: createUser
            requestBody:
                description: Datos del usuario
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UsuarioCrearRequest"
            responses:
                "200":
                    description: Operacion exitosa
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/RespuestaComun"
                "400":
                    description: Datos inconrrectos
                "404":
                    description: Operacion no encontrada
            security:
                - bearerAuth: []
    /usuarios/{usuarioId}:
        get:
            tags:
                - Usuarios
            summary: Obtener un usuario por id
            description: Obtiene usuario en especifico
            operationId: getUserById
            parameters:
                - name: usuarioId
                  in: path
                  description: ID de usuario
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    description: Operacion exitosa
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/RespuestaComun"
                                    - properties:
                                          data:
                                              $ref: "#/components/schemas/UsuarioRespuesta"
                "404":
                    description: Operacion no encontrada
            security:
                - bearerAuth: []
        put:
            tags:
                - Usuarios
            summary: Actualiza un usuario por id
            description: Actualiza usuario en especifico
            operationId: updateUserById
            parameters:
                - name: usuarioId
                  in: path
                  description: ID de usuario
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: Datos del usuario
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UsuarioUpdateRequest"
            responses:
                "200":
                    description: Operacion exitosa
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: "#/components/schemas/RespuestaComun"
                                    - properties:
                                          data:
                                              $ref: "#/components/schemas/UsuarioRespuesta"
                "404":
                    description: Operacion no encontrada
            security:
                - bearerAuth: []
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        UsuarioCrearRequest:
            type: object
            required:
                [
                    nombre,
                    apellido,
                    rol,
                    telefono_personal,
                    password,
                ]
            properties:
                email:
                    type: string
                nombre:
                    type: string
                apellido:
                    type: string
                rol:
                    type: integer
                    format: int32
                avatar:
                    type: string
                telefono_personal:
                    type: integer
                    format: int32
                password:
                    type: string
                negocio_id:
                    type: integer
                    format: int32
        UsuarioUpdateRequest:
            type: object
            required:
                [
                    nombre,
                    apellido,
                    rol,
                    telefono_personal,
                    password,
                ]
            properties:
                nombre:
                    type: string
                apellido:
                    type: string
                rol:
                    type: integer
                    format: int32
                avatar:
                    type: string
                telefono_personal:
                    type: integer
                    format: int32
        UsuariosRespuesta:
            type: array
            items:
                $ref: "#/components/schemas/UsuarioRespuesta"
        UsuarioRespuesta:
            type: object
            properties:
                email:
                    type: string
                nombre:
                    type: string
                apellido:
                    type: string
                rol:
                    type: integer
                    format: int32
                avatar:
                    type: string
                telefono_personal:
                    type: integer
                    format: int32
                negocio:
                    $ref: "#/components/schemas/NegocioRespuesta"
...
        NegocioRespuesta:
            type: object
            properties:
                nombre:
                    type: string
                categoria_id:
                    type: integer
                    format: int32
                domicilio:
                    type: string
                logo:
                    type: string
                fachada:
                    type: string
                codigo_recomendacion:
                    type: string
                telefono_negocio:
                    type: integer
                    format: int32
                calificacion:
                    type: integer
                    format: int32
...
        RespuestaComun:
            type: object
            properties:
                codigo:
                    type: integer
                    format: int32
                message:
                    type: string